% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proName}
\alias{proName}
\title{Convert names of provinces between different formats.}
\usage{
proName(
  name,
  to = "eng",
  from = NULL,
  asCharacter = TRUE,
  caseSensitive = FALSE,
  NAifNotFound = FALSE,
  fuzzy = FALSE,
  asTibble = FALSE
)
}
\arguments{
\item{name}{character, name of the province}

\item{to}{type of name to return (column names of \code{CEPRE:::.proNames})}

\item{from}{character or NULL, the type of name to convert from (if known)}

\item{asCharacter}{logical,}

\item{caseSensitive}{logical}

\item{NAifNotFound}{logical,}

\item{fuzzy}{ignored}

\item{asTibble}{logical, FALSE by default}
}
\description{
Convert names of provinces between different formats.
}
